# docs/swagger.yaml
definitions:
  domain.BmiRequest:
    properties:
      height:
        type: number
      name:
        type: string
      weight:
        type: number
    required:
    - height
    - name
    - weight
    type: object
  domain.BmiResponse:
    properties:
      bmi:
        type: number
      created_at:
        type: string
      height:
        type: number
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      weight:
        type: number
    type: object
  rest.ResponseError:
    properties:
      message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/bmi:
    get:
      description: Get BMIs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BmiResponse'
            type: array
      summary: Get BMIs
    post:
      consumes:
      - application/json
      description: Create BMI
      parameters:
      - description: BMI request object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BmiRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.BmiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Create BMI
  /api/bmi/{id}:
    delete:
      description: Delete BMI by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Delete BMI by ID
    get:
      description: Get BMI by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BmiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Get BMI by ID
    put:
      consumes:
      - application/json
      description: Update BMI by ID
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: BMI request object
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BmiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BmiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Update BMI by ID
swagger: "2.0"